Add files via upload
The hand module is the Complete Module that you can add to your program, and the number, Arduino, and the Volume control files include the src code of the individual code for each of the mini-projects you can implement using the HandModule, the number program is not get been implemented in the Handmodule, but it will be in the near future. Another future mini-project that will be coming is a rock paper scissors game against an ai.